workflows: increase tests timeout up to 25 minutes
I'm tired of tests failing due to timeout, GitHub machines are too slow.
I propose to use 25m for some time and see if we're still having
timeouting tests. If so, then we need to refactor timeouting tests,
because failure may be caused by inappropriate test design. If not, then
we may try to reduce tests timeout to 20m later.

Close #3367 along the way.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>