network: take care of TestServerStartAndShutdown
We don't have a reliable way to know when transports are started since
their start is being performed in a separate goroutine:

https://github.com/nspcc-dev/neo-go/blob/927dbb6dc48d578ae731e5a067c82d0611642338/pkg/network/server.go#L297-L299

And transports start is not connected with main server routine, thus,
just wait for some time for the transports goroutine to be started.

Also wait for the peer to be properly registered.

Close #3399.

Signed-off-by: Anna Shaleva <shaleva.ann@nspcc.ru>