config: drop MTB value for NeoFS networks to three days
We want to cut the tail more aggressively, what matters is state, but not
old blocks and transactions that contracts don't use in any way. #3493
suggests that it's safe to limit the tail to one day, but let's be a bit
more conservative for now.

As a sidenote, EVM only allows to fetch things from the recent 256 blocks.

Signed-off-by: Roman Khimov <roman@nspcc.ru>